Finished refactoring leaderboards panel
TODO: hook the header map info collection to the site, hook statistics gathering to the site
