+++ mp/game/momentum/resource/ui/leaderboards/times.res
@@ -1,0 +1,303 @@
+"resource/ui/leaderboards/times.res"
+{
+	"OnlineTimesStatus"
+	{
+		"ControlName"	"Label"
+		"fieldName"		"OnlineTimesStatus"
+		"xpos"			"0"
+		"ypos"			"-110"
+		"wide"			"0"
+		"tall"			"0"
+		"autoResize"	"0"
+		"pinCorner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"labelText"		"#MOM_API_WaitingForResponse"
+		"textAlignment"	"center"
+		"dulltext"		"0"
+		"brighttext"	"1"
+		"font"			"BigNoodle"
+		"wrap"			"0"
+		"centerwrap"	"0"
+		"textinsetx"	"0"
+		"textinsety"	"0"
+		"auto_wide_tocontents" "1"
+        "auto_tall_tocontents" "1"
+        "zpos" "20"
+        "proportionalToParent" "1"
+        "pin_to_sibling" "Top10Leaderboards"
+        "pin_to_sibling_corner" "6"
+        "pin_corner_to_sibling" "6"
+	}
+    "Top10Leaderboards"
+	{
+		"ControlName"	"SectionedListPanel"
+		"fieldName"		"OnlineLeaderboards"
+		"xpos"			"3"
+		"ypos"			"27"
+		"wide"			"468"
+		"tall"			"240"
+		"autoResize"	"0"
+		"pinCorner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"tabPosition"	"2"
+        "actionsignallevel" "1"
+        "paintborder" "1"
+        "keyboardinputenabled" "1"
+        "mouseinputenabled" "1"
+        "vertical_scrollbar" "0" // MOM_TODO: Do we want the player to be able to explore more?
+        //"linespacing" "32"
+	}
+    "AroundLeaderboards"
+	{
+		"ControlName"	"SectionedListPanel"
+		"fieldName"		"AroundLeaderboards"
+		"xpos"			"3"
+		"ypos"			"27"
+		"wide"			"468"
+		"tall"			"240"
+		"autoResize"	"0"
+		"pinCorner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"tabPosition"	"2"
+        "actionsignallevel" "1"
+        "paintborder" "1"
+        "keyboardinputenabled" "1"
+        "mouseinputenabled" "1"
+        "vertical_scrollbar" "0" // MOM_TODO: Do we want the player to be able to explore more?
+        //"linespacing" "32"
+	}
+	"LocalLeaderboards"
+	{
+		"ControlName"	"SectionedListPanel"
+		"fieldName"		"LocalLeaderboards"
+		"xpos"			"3"
+		"ypos"			"27"
+		"wide"			"468"
+		"tall"			"240"
+		"autoResize"	"0"
+		"pinCorner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"	"3"
+        "actionsignallevel" "1"
+        "paintborder" "1"
+        "keyboardinputenabled" "1"
+        "mouseinputenabled" "1"
+        "vertical_scrollbar" "0" // MOM_TODO: Do we want the player to be able to explore past runs?
+	}	
+	"FriendsLeaderboards"
+	{
+		"ControlName"	"SectionedListPanel"
+		"fieldName"		"FriendsLeaderboards"
+		"xpos"			"3"
+		"ypos"			"27"
+		"wide"			"468"
+		"tall"			"240"
+		"autoResize"	"0"
+		"pinCorner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"tabPosition"	"2"
+        "actionsignallevel" "1"
+        "paintborder" "1"
+        "keyboardinputenabled" "1"
+        "mouseinputenabled" "1"
+        "vertical_scrollbar" "0" // MOM_TODO: Do we want the player to be able to explore more?
+        //"linespacing" "32"
+	}
+    "LocalLeaderboardsButton"
+    {
+        "ControlName" "Button"
+        "fieldName" "LocalLeaderboardsButton"
+        "xpos" "8"
+        "ypos" "2"
+        "wide"		"88"
+		"tall"		"16"
+		"autoResize"		"0"
+		"pinCorner"		"2"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"0"
+		"tabPosition"		"0"
+		"labelText"		"#MOM_Leaderboards_Local"
+		"textAlignment"		"center"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+        "command" "ShowLocal"
+		"centerwrap"		"0"
+		"textinsetx"		"6"
+		"textinsety"		"0"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+		"Font"			"BigNoodleSmall"
+        "actionsignallevel" "1"
+        "mouseinputenabled" "1"
+    }
+	"GlobalLeaderboardsButton"
+	{
+		"ControlName"		"Button"
+		"fieldName"		"GlobalLeaderboardsButton"
+		"xpos"		"41"
+		"ypos"		"2"
+		"wide"		"88"
+		"tall"		"16"
+		"autoResize"		"0"
+		"pinCorner"		"2"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"labelText"		"#MOM_Leaderboards_Global"
+		"textAlignment"		"center"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+        "command" "ShowGlobal"
+		"centerwrap"		"0"
+		"textinsetx"		"6"
+		"textinsety"		"0"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+		"Font"			"BigNoodleSmall"
+        "actionsignallevel" "1"
+        "mouseinputenabled" "1"
+	}
+	"GlobalTop10Button"
+	{
+		"ControlName"		"Button"
+		"fieldName"		"GlobalTop10Button"
+		"xpos"		"81"
+		"ypos"		"2"
+		"wide"		"88"
+		"tall"		"16"
+		"autoResize"		"0"
+		"pinCorner"		"2"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"0"
+		"enabled"		"0"
+		"tabPosition"		"0"
+		"labelText"		"#MOM_Leaderboards_Top10"
+		"textAlignment"		"center"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+        "command" "GlobalTypeTop10"
+		"centerwrap"		"0"
+		"textinsetx"		"6"
+		"textinsety"		"0"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+		"Font"			"BigNoodleSmall"
+        "actionsignallevel" "1"
+        "mouseinputenabled" "1"
+	}
+	"GlobalAroundButton"
+	{
+		"ControlName"		"Button"
+		"fieldName"		"GlobalAroundButton"
+		"xpos"		"116"
+		"ypos"		"2"
+		"wide"		"88"
+		"tall"		"16"
+		"autoResize"		"0"
+		"pinCorner"		"2"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"labelText"		"#MOM_Leaderboards_Around"
+		"textAlignment"		"center"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+		"centerwrap"		"0"
+		"textinsetx"		"6"
+		"textinsety"		"0"
+        "command" "GlobalTypeAround"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+		"Font"			"BigNoodleSmall"
+        "actionsignallevel" "1"
+        "mouseinputenabled" "1"
+	}
+	"FriendsLeaderboardsButton"
+	{
+		"ControlName"		"Button"
+		"fieldName"		"FriendsLeaderboardsButton"
+		"xpos"		"156"
+		"ypos"		"2"
+		"wide"		"88"
+		"tall"		"16"
+		"autoResize"		"0"
+		"pinCorner"		"2"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"labelText"		"#MOM_Leaderboards_Friends"
+		"textAlignment"		"center"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+        "command" "ShowFriends"
+		"centerwrap"		"0"
+		"textinsetx"		"6"
+		"textinsety"		"0"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+		"Font"			"BigNoodleSmall"
+        "actionsignallevel" "1"
+        "mouseinputenabled" "1"
+	}
+    "FilterButton"
+    {
+        "ControlName" "ToggleButton"
+        "fieldName" "FilterButton"
+        "xpos"		"435"
+		"ypos"		"2"
+		"wide"		"90"
+		"tall"		"16"
+		"autoResize"		"0"
+		"pinCorner"		"2"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"labelText"		"#MOM_Leaderboards_Filter"
+		"textAlignment"		"center"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+        "command" "ShowFilter"
+		"centerwrap"		"0"
+		"textinsetx"		"6"
+		"textinsety"		"0"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+		"Font"			"BigNoodleSmall"
+        "actionsignallevel" "1"
+        "mouseinputenabled" "1"
+    }
+}
